feature: sidebar component made
